+++ src/Modules/Store/StoreControl.php
@@ -132,6 +132,7 @@ class StoreControl extends Control
 					'added' => $store['added'],
 					'region' => $store['bezirk_name'],
 					'address' => $store['anschrift'],
+					'city' => $store['stadt'],
 				];
 			}, $stores);